🐛 fix(config): widen confirm_countdown_secs from u8 to u32
Addresses Copilot review on PR #66 (src/config.rs:147). The PR
documented "values above 5 are clamped on read", but a `u8` field
fails TOML deserialization at parse time for any value above 255 —
turning a typo like `confirm_countdown_secs = 300` into a hard
`Config::load_for_repo` error instead of the documented clamp-to-5.

Widening to `u32` keeps the field cheap (4 bytes), preserves
ergonomic round-tripping for arbitrarily-large typoed values, and
lets `effective_confirm_countdown_secs()` enforce the documented
range cap on every read.

The `as u64` cast on the consumer side is replaced with an
infallible `u64::from()` since `u32 → u64` is a widening conversion.

Pinned by tests/config_tests.rs::tui_countdown_value_above_u8_max_still_clamps
(red against the previous `u8` field — fails at TOML parse, not at
the assert — green after this change).

Addresses review by Copilot on PR #66.

+++ src/config.rs
@@ -136,15 +136,19 @@ fn default_trunks() -> Vec<String> {
 /// destructive action actually fires. `0` disables the countdown and
 /// falls back to the classic single-keystroke confirm even when delete-
 /// branch is armed; the value is clamped to `5` at read time so a typo
-/// like `confirm_countdown_secs = 30` can never strand a destructive
-/// path behind a 30-second wait.
+/// like `confirm_countdown_secs = 300` can never strand a destructive
+/// path behind a 300-second wait.
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct TuiConfig {
   /// Safety countdown (in seconds) applied to the confirm overlay when
-  /// `delete_branch_on_remove` is ON. Accepts `0..=5`; values outside
-  /// the range are clamped on read via [`Self::effective_confirm_countdown_secs`].
+  /// `delete_branch_on_remove` is ON. Accepts any non-negative integer;
+  /// values above [`Self::MAX_CONFIRM_COUNTDOWN_SECS`] are clamped on
+  /// read via [`Self::effective_confirm_countdown_secs`]. The field is
+  /// `u32` (rather than `u8`) so a typo like `confirm_countdown_secs = 300`
+  /// still round-trips through TOML deserialization and reaches the
+  /// clamp instead of erroring out at parse time.
   #[serde(default = "default_confirm_countdown_secs")]
-  pub confirm_countdown_secs: u8,
+  pub confirm_countdown_secs: u32,
 }
 
 impl Default for TuiConfig {
@@ -158,18 +162,18 @@ impl Default for TuiConfig {
 impl TuiConfig {
   /// Documented range cap. Centralised so the TUI and the doctor share
   /// the same clamp logic.
-  pub const MAX_CONFIRM_COUNTDOWN_SECS: u8 = 5;
+  pub const MAX_CONFIRM_COUNTDOWN_SECS: u32 = 5;
 
   /// Effective countdown value used by the TUI, clamped to
   /// `[0, MAX_CONFIRM_COUNTDOWN_SECS]`. The raw field stays at the
   /// user's value so a future doctor check can surface "your config
   /// asked for X but we capped at 5".
-  pub fn effective_confirm_countdown_secs(&self) -> u8 {
+  pub fn effective_confirm_countdown_secs(&self) -> u32 {
     self.confirm_countdown_secs.min(Self::MAX_CONFIRM_COUNTDOWN_SECS)
   }
 }
 
-fn default_confirm_countdown_secs() -> u8 {
+fn default_confirm_countdown_secs() -> u32 {
   3
 }
